+++ watch/cli/smartflow/run_sc_fusion.py
@@ -29,10 +29,6 @@ class SCFusionConfig(scfg.DataConfig):
             '''))
     output_path = scfg.Value(None, type=str, position=3, required=True, help='S3 path for output JSON')
 
-    sc_track_fn = scfg.Value('class_heatmaps', type=str, help=ub.paragraph(
-            '''
-            Tracking function to use for generating sites
-            '''))
     aws_profile = scfg.Value(None, type=str, help=ub.paragraph(
             '''
             AWS Profile to use for AWS S3 CLI commands
@@ -71,28 +67,57 @@ def main():
 def run_sc_fusion_for_baseline(config):
     from watch.cli.smartflow_ingress import smartflow_ingress
     from watch.cli.smartflow_egress import smartflow_egress
-    from watch.tasks.fusion.predict import predict
+    from watch.tasks.fusion.predict import predict  # NOQA
     from watch.tasks.fusion.datamodules.temporal_sampling import TimeSampleError
     from watch.utils.util_framework import download_region, determine_region_id
     from kwutil.util_yaml import Yaml
     from watch.utils import util_framework
+    from watch.mlops import smart_pipeline
 
     if config.aws_profile is not None:
         # This should be sufficient, but it is not tested.
         from watch.utils import util_fsspec
         util_fsspec.S3Path._new_fs(profile=config.aws_profile)
 
+    ####
+    # DEBUGGING:
+    # Print info about what version of the code we are running on
+    import watch
+    print('Print current version of the code')
+    ub.cmd('git log -n 1', verbose=3, cwd=ub.Path(watch.__file__).parent)
+
     # 1. Ingress data
     print("* Running baseline framework kwcoco ingress *")
+
+    # ingress_dir = ub.Path('/home/joncrall/data/dvc-repos/smart_expt_dvc/_airflow/temp').ensuredir()
     ingress_dir = ub.Path('/tmp/ingress')
+
     ingressed_assets = smartflow_ingress(
-        config.input_path,
-        ['cropped_region_models_bas',
-         'cropped_kwcoco_for_sc',
-         'cropped_kwcoco_for_sc_assets'],
-        ingress_dir,
-        config.aws_profile,
-        config.dryrun)
+        input_path=config.input_path,
+        assets=[
+            {'key': 'cropped_region_models_bas'},
+            {'key': 'sv_out_region_models', 'allow_missing': False},
+            {'key': 'cropped_kwcoco_for_sc'},
+            {'key': 'cropped_kwcoco_for_sc_assets'}
+        ],
+        outdir=ingress_dir,
+        aws_profile=config.aws_profile,
+        dryrun=config.dryrun
+    )
+
+    # Get the first set of BAS site summaries that are available
+    region_model_key_priority = [
+        # Use filtered SV site summaries when possible
+        # Otherwise fallback to bas site summaries
+        'sv_out_region_models',
+        'cropped_region_models_bas',
+    ]
+    for key in region_model_key_priority:
+        input_site_summary_dpath = ingressed_assets[key]
+        if os.path.exists(input_site_summary_dpath):
+            break
+    assert os.path.exists(input_site_summary_dpath)
+    print(f'Found input site summary dpath: {input_site_summary_dpath}')
 
     # # 2. Download and prune region file
     print("* Downloading and pruning region file *")
@@ -109,20 +134,25 @@ def run_sc_fusion_for_baseline(config):
     region_id = determine_region_id(local_region_path)
 
     sc_fusion_kwcoco_path = ingress_dir / 'sc_fusion_kwcoco.json'
-    cropped_region_models_bas = ingressed_assets['cropped_region_models_bas']
 
-    site_models_outdir = ingress_dir / 'sc_out_site_models'
-    os.makedirs(site_models_outdir, exist_ok=True)
-    region_models_outdir = ingress_dir / 'sc_out_region_models'
-    os.makedirs(region_models_outdir, exist_ok=True)
+    tracked_sc_kwcoco_path = '_tracked'.join(
+        os.path.splitext(sc_fusion_kwcoco_path))
 
-    site_models_manifest_outdir = ingress_dir / 'tracking_manifests_sc'
-    os.makedirs(site_models_manifest_outdir, exist_ok=True)
+    site_models_outdir = (ingress_dir / 'sc_out_site_models').ensuredir()
+    region_models_outdir = (ingress_dir / 'sc_out_region_models').ensuredir()
+    site_models_manifest_outdir = (ingress_dir / 'tracking_manifests_sc').ensuredir()
+
     site_models_manifest_outpath = site_models_manifest_outdir / 'site_models_manifest.json'
     # Copy input region model into region_models outdir to be updated
     # (rather than generated from tracking, which may not have the
     # same bounds as the original)
     shutil.copy(local_region_path, region_models_outdir / f'{region_id}.geojson')
+
+    region_models_manifest_fpath = ingress_dir / 'sc_out_region_models_manifest.json'
+
+    print('* Printing current directory contents (1/5)')
+    cwd_paths = sorted([p.resolve() for p in ingress_dir.glob('*')])
+    print('cwd_paths = {}'.format(ub.urepr(cwd_paths, nl=1)))
 
     # 3.1. Check that we have at least one "video" (BAS identified
     # site) to run over; if not skip SC fusion and KWCOCO to GeoJSON
@@ -134,84 +164,66 @@ def run_sc_fusion_for_baseline(config):
         print('*********************')
         print("* Running SC fusion *")
 
-        # TODO: remove these defaults or replace them with whatever is the
-        # default in predict. The params should be fully given in the DAG, not
-        # here.
-        default_sc_pxl_config = ub.udict({
-            # 'tta_fliprot': 0.0,
-            # 'tta_time': 0.0,
-            # 'chip_overlap': 0.3,
-            # 'input_space_scale': '8GSD',
-            # 'window_space_scale': '8GSD',
-            # 'output_space_scale': '8GSD',
-            # 'time_span': '6m',
-            # 'time_sampling': 'auto',
-            # 'time_steps': '12',
-            # 'chip_dims': 'auto',
-            # 'set_cover_algo': None,
-            # 'resample_invalid_frames': 3,
-            # 'observable_threshold': 0.0,
-            # 'mask_low_quality': True,
-            # 'drop_unused_frames': True,
-            # 'num_workers': 2,
-            # 'batch_size': 1,
-            # 'write_workers': 0,
-            # 'package_fpath': None,
-        })
-
-        sc_pxl_config = default_sc_pxl_config | Yaml.coerce(config.sc_pxl_config or {})
-
+        # The params should be fully given in the DAG.
+        sc_pxl_config = Yaml.coerce(config.sc_pxl_config or {})
         if sc_pxl_config.get('package_fpath', None) is None:
             raise ValueError('Requires package_fpath')
 
+        sc_pxl = smart_pipeline.SC_HeatmapPrediction(root_dpath=ingress_dir)
+        sc_pxl.configure({
+            'pred_pxl_fpath': sc_fusion_kwcoco_path,
+            'test_dataset': ingressed_assets['cropped_kwcoco_for_sc'],
+        } | sc_pxl_config)
+        command = sc_pxl.command()
+
         try:
-            predict(devices='0,',
-                    write_preds=False,
-                    write_probs=True,
-                    with_change=False,
-                    with_saliency=False,
-                    with_class=True,
-                    test_dataset=ingressed_assets['cropped_kwcoco_for_sc'],
-                    pred_dataset=sc_fusion_kwcoco_path,
-                    **sc_pxl_config)
+            ub.cmd(command, check=True, verbose=3, system=True)
+            print('* Printing current directory contents (2/5)')
+            cwd_paths = sorted([p.resolve() for p in ingress_dir.glob('*')])
+            print('cwd_paths = {}'.format(ub.urepr(cwd_paths, nl=1)))
         except TimeSampleError:
+            # FIXME: wont work anymore with mlops. Not sure if needed.
+            # Can always catch a CalledProcessError and inspect stdout
             print("* Error with time sampling during SC Predict "
                   "(shown below) -- attempting to continue anyway")
             traceback.print_exception(*sys.exc_info())
         else:
+
             # 4. Compute tracks (SC)
             print('*************************')
             print("* Computing tracks (SC) *")
 
-            # TODO: remove these defaults or replace them with whatever is the
-            # default in tracker. The params should be fully given in the DAG,
-            # not here.
-            default_sc_track_kwargs = ub.udict({
-                "boundaries_as": "polys",
-                "resolution": 8,
-                "min_area_square_meters": 7200,
-                "thresh": 0.07,
-            })
-            sc_track_kwargs = default_sc_track_kwargs | Yaml.coerce(config.sc_poly_config or {})
-
+            # Params are fully specified in the DAG
+            sc_track_kwargs = Yaml.coerce(config.sc_poly_config or {})
             tracked_sc_kwcoco_path = '_tracked'.join(
                 os.path.splitext(sc_fusion_kwcoco_path))
-            ub.cmd([
-                'python', '-m', 'watch.cli.run_tracker',
-                '--input_kwcoco', sc_fusion_kwcoco_path,
-                '--out_site_summaries_dir', region_models_outdir,
-                '--out_sites_dir', site_models_outdir,
-                '--out_sites_fpath', site_models_manifest_outpath,
-                '--out_kwcoco', tracked_sc_kwcoco_path,
-                '--default_track_fn', config.sc_track_fn,
-                '--site_summary', ub.Path(cropped_region_models_bas) / '*.geojson',
-                '--append_mode', 'True',
-                '--track_kwargs', json.dumps(sc_track_kwargs)],
-                check=True, verbose=3, capture=False)
+            final_sc_poly_config = {
+                'pred_pxl_fpath': sc_fusion_kwcoco_path,               # Sets --input_kwcoco
+                'site_summaries_fpath': region_models_manifest_fpath,  # Sets --out_site_summaries_fpath
+                'site_summaries_dpath': region_models_outdir,          # Sets --out_site_summaries_dir
+                'sites_dpath': site_models_outdir,                     # Sets --out_sites_dir
+                'sites_fpath': site_models_manifest_outpath,           # Sets --out_sites_fpath
+                'poly_kwcoco_fpath': tracked_sc_kwcoco_path,           # Sets --out_kwcoco
+                'site_summary': ub.Path(input_site_summary_dpath) / '*.geojson',  # Sets --site_summary
+                'append_mode': True,
+            } | sc_track_kwargs
+            sc_poly = smart_pipeline.SC_PolygonPrediction(root_dpath=ingress_dir)
+            sc_poly.configure(final_sc_poly_config)
+            command = sc_poly.command()
+            ub.cmd(command, check=True, verbose=3, system=True)
+
+            print('* Printing current directory contents (3/5)')
+            cwd_paths = sorted([p.resolve() for p in ingress_dir.glob('*')])
+            print('cwd_paths = {}'.format(ub.urepr(cwd_paths, nl=1)))
+
+            # Add in intermediate outputs for debugging
+            ingressed_assets['sc_heatmap_kwcoco_file'] = sc_fusion_kwcoco_path
+            ingressed_assets['sc_tracked_kwcoco_file'] = tracked_sc_kwcoco_path
 
     cropped_site_models_outdir = ingress_dir / 'cropped_site_models'
     os.makedirs(cropped_site_models_outdir, exist_ok=True)
     cropped_region_models_outdir = ingress_dir / 'cropped_region_models'
+    sc_heatmap_dpath = ingress_dir / '_assets'
     os.makedirs(cropped_region_models_outdir, exist_ok=True)
 
     ub.cmd([
@@ -222,12 +234,20 @@ def run_sc_fusion_for_baseline(config):
         '--new_region_dpath', cropped_region_models_outdir
     ], check=True, verbose=3, capture=False)
 
+    print('* Printing current directory contents (4/5)')
+    cwd_paths = sorted([p.resolve() for p in ingress_dir.glob('*')])
+    print('cwd_paths = {}'.format(ub.urepr(cwd_paths, nl=1)))
+
     # Validate and fix all outputs
     print('Fixup and validate outputs')
     util_framework.fixup_and_validate_site_and_region_models(
         region_dpath=cropped_region_models_outdir,
         site_dpath=cropped_site_models_outdir,
     )
+
+    print('* Printing current directory contents (5/5)')
+    cwd_paths = sorted([p.resolve() for p in ingress_dir.glob('*')])
+    print('cwd_paths = {}'.format(ub.urepr(cwd_paths, nl=1)))
 
     # 5. Egress (envelop KWCOCO dataset in a STAC item and egress;
     #    will need to recursive copy the kwcoco output directory up to
@@ -237,8 +257,14 @@ def run_sc_fusion_for_baseline(config):
     ingressed_assets['cropped_region_models_sc'] = cropped_region_models_outdir
 
     # Add in intermediate outputs for debugging
-    ingressed_assets['sc_heatmap_kwcoco_file'] = sc_fusion_kwcoco_path
-    ingressed_assets['sc_tracked_kwcoco_file'] = tracked_sc_kwcoco_path
+    EGRESS_FUSION_HEATMAPS = True
+    if EGRESS_FUSION_HEATMAPS:
+        ingressed_assets['sc_heatmap_kwcoco_file'] = sc_fusion_kwcoco_path
+        ingressed_assets['sc_tracked_kwcoco_file'] = tracked_sc_kwcoco_path
+        ingressed_assets['sc_heatmap_assets'] = sc_heatmap_dpath
+        ingressed_assets['sc_tracking_manifest_dpath'] = site_models_manifest_outdir
+        if region_models_manifest_fpath.exists():
+            ingressed_assets['sc_tracking_manifest_fpath'] = region_models_manifest_fpath
 
     smartflow_egress(ingressed_assets,
                      local_region_path,